fixed rust file size
